Refactorice change listEvents to public folder
