+++ galeria/src/pages/dynamic/styles/formstyles/FormStyles.css
@@ -4,6 +4,10 @@
   flex-direction: column;
 }
 
+#form-container-events h2 {
+  padding: 10px 30px;
+}
+
 input {
   padding: 15px;
   font-size: 16px;
@@ -12,17 +16,15 @@ input {
 }
 
 textarea {
-    background-color: whitesmoke;
-    border: none;
-    padding: 10px;
-    width: 300px;
+  background-color: whitesmoke;
+  border: none;
+  padding: 10px;
+  width: 300px;
 }
-
 
 input:not(.number-input, .radio-input) {
   width: 300px;
 }
-
 
 #date-input {
   font-family: sans-serif;
